fix lambda docker
